refactor: move stateless change helpers out of SubjectTransaction
Move the pure change-collection helpers (RevertInProcess, IntersectByProperty,
ExcludeByProperty, DetectConflicts, Concat) out of the oversized
SubjectTransaction into the change-operations class, and rename that class from
SubjectPropertyChangeExtensions to SubjectPropertyChangeOperations to reflect
that it now holds apply, rollback, partition and conflict operations. The
transaction keeps only commit/lifecycle logic and the transaction-semantic
wrappers (ThrowIfConflictsDetected, CreateFailureException). No behavior change.
